+++ library/LibUtilities/TimeIntegration/TimeIntegrationAlgorithmGLM.cpp
@@ -1058,7 +1058,6 @@ std::ostream &operator<<(std::ostream &os,
                 os << std::endl;
             }
 
-            size_t imexflag = (type == eIMEX) ? 2 : 1;
             for (size_t i = 0;
                  i < (r + imexflag * s) * (oswidth + 1) + imexflag * 2 - 1; i++)
             {